+++ test/manifest.ts
@@ -23,6 +23,7 @@ import {
   stubSuggesterWithSnapshot,
   assertHasUpdate,
   dateSafe,
+  safeSnapshot,
 } from './helpers';
 import {expect} from 'chai';
 import {Version} from '../src/version';
@@ -226,6 +227,67 @@ describe('Manifest', () => {
         manifest.repositoryConfig['packages/cron-utils'].pullRequestTitlePattern
       ).to.eql('chore${scope}: send it v${version}');
     });
+
+    it('should read custom tag separator from manifest', async () => {
+      const getFileContentsStub = sandbox.stub(
+        github,
+        'getFileContentsOnBranch'
+      );
+      getFileContentsStub
+        .withArgs('release-please-config.json', 'main')
+        .resolves(
+          buildGitHubFileContent(
+            fixturesPath,
+            'manifest/config/tag-separator.json'
+          )
+        )
+        .withArgs('.release-please-manifest.json', 'main')
+        .resolves(
+          buildGitHubFileContent(
+            fixturesPath,
+            'manifest/versions/versions.json'
+          )
+        );
+      const manifest = await Manifest.fromManifest(
+        github,
+        github.repository.defaultBranch
+      );
+      expect(manifest.repositoryConfig['.'].tagSeparator).to.eql('-');
+      expect(
+        manifest.repositoryConfig['packages/bot-config-utils'].tagSeparator
+      ).to.eql('/');
+    });
+
+    it('should read custom include component in tag from manifest', async () => {
+      const getFileContentsStub = sandbox.stub(
+        github,
+        'getFileContentsOnBranch'
+      );
+      getFileContentsStub
+        .withArgs('release-please-config.json', 'main')
+        .resolves(
+          buildGitHubFileContent(
+            fixturesPath,
+            'manifest/config/include-component-in-tag.json'
+          )
+        )
+        .withArgs('.release-please-manifest.json', 'main')
+        .resolves(
+          buildGitHubFileContent(
+            fixturesPath,
+            'manifest/versions/versions.json'
+          )
+        );
+      const manifest = await Manifest.fromManifest(
+        github,
+        github.repository.defaultBranch
+      );
+      expect(manifest.repositoryConfig['.'].includeComponentInTag).to.be.false;
+      expect(
+        manifest.repositoryConfig['packages/bot-config-utils']
+          .includeComponentInTag
+      ).to.be.true;
+    });
   });
 
   describe('fromConfig', () => {
@@ -244,6 +306,13 @@ describe('Manifest', () => {
             labels: [],
             files: [],
           },
+        },
+      ]);
+      mockReleases(github, [
+        {
+          tagName: 'v1.2.3',
+          sha: 'abc123',
+          url: 'http://path/to/release',
         },
       ]);
 
@@ -270,6 +339,13 @@ describe('Manifest', () => {
             labels: [],
             files: [],
           },
+        },
+      ]);
+      mockReleases(github, [
+        {
+          tagName: 'v1.2.3',
+          sha: 'abc123',
+          url: 'http://path/to/release',
         },
       ]);
 
@@ -301,6 +377,13 @@ describe('Manifest', () => {
           },
         },
       ]);
+      mockReleases(github, [
+        {
+          tagName: 'v1.2.3',
+          sha: 'abc123',
+          url: 'http://path/to/release',
+        },
+      ]);
 
       const manifest = await Manifest.fromConfig(github, 'target-branch', {
         releaseType: 'simple',
@@ -327,6 +410,13 @@ describe('Manifest', () => {
             labels: [],
             files: [],
           },
+        },
+      ]);
+      mockReleases(github, [
+        {
+          tagName: 'foobar-v1.2.3',
+          sha: 'abc123',
+          url: 'http://path/to/release',
         },
       ]);
 
@@ -578,6 +668,61 @@ describe('Manifest', () => {
         '3.3.3'
       );
     });
+    it('finds manually tagged release commit over earlier automated commit', async () => {
+      mockCommits(github, [
+        {
+          sha: 'abc123',
+          message: 'some commit message',
+          files: [],
+        },
+        {
+          sha: 'def234',
+          message: 'this commit should be found',
+          files: [],
+        },
+        {
+          sha: 'ghi345',
+          message: 'some commit message',
+          files: [],
+          pullRequest: {
+            title: 'chore: release 3.3.1',
+            headBranchName: 'release-please/branches/main',
+            baseBranchName: 'main',
+            number: 123,
+            body: '',
+            labels: [],
+            files: [],
+          },
+        },
+      ]);
+      mockReleases(github, [
+        {
+          tagName: 'v3.3.2',
+          sha: 'def234',
+          url: 'http://path/to/release',
+        },
+        {
+          tagName: 'v3.3.1',
+          sha: 'ghi345',
+          url: 'http://path/to/release',
+        },
+      ]);
+      mockTags(github, []);
+
+      const manifest = await Manifest.fromConfig(github, 'target-branch', {
+        releaseType: 'simple',
+        bumpMinorPreMajor: true,
+        bumpPatchForMinorPreMajor: true,
+      });
+      expect(Object.keys(manifest.repositoryConfig)).lengthOf(1);
+      expect(
+        Object.keys(manifest.releasedVersions),
+        'found release versions'
+      ).lengthOf(1);
+      expect(Object.values(manifest.releasedVersions)[0].toString()).to.eql(
+        '3.3.2'
+      );
+    });
   });
 
   describe('buildPullRequests', () => {
@@ -641,6 +786,7 @@ describe('Manifest', () => {
       });
 
       it('should honour the manifestFile argument in Manifest.fromManifest', async () => {
+        mockTags(github, []);
         const getFileContentsStub = sandbox.stub(
           github,
           'getFileContentsOnBranch'
@@ -1194,6 +1340,7 @@ describe('Manifest', () => {
           files: ['path/b/foo'],
         },
       ]);
+      mockTags(github, []);
       const config = {
         'bootstrap-sha': 'cccccc',
         'separate-pull-requests': true,
@@ -1284,6 +1431,7 @@ describe('Manifest', () => {
           },
         },
       ]);
+      mockTags(github, []);
       const config = {
         'last-release-sha': 'bbbbbb',
         'separate-pull-requests': true,
@@ -1531,6 +1679,7 @@ describe('Manifest', () => {
           files: ['path/a/foo'],
         },
       ]);
+      mockTags(github, []);
       const config = {
         packages: {
           'path/a': {
@@ -1629,6 +1778,7 @@ describe('Manifest', () => {
 
     it('should handle extra files', async () => {
       mockReleases(github, []);
+      mockTags(github, []);
       mockCommits(github, [
         {
           sha: 'aaaaaa',
@@ -1690,6 +1840,67 @@ describe('Manifest', () => {
           'pkg/b/bbb.properties', // should be at root
           'pkg/c/pkg-c.properties', // should be up one level
         ]);
+    });
+
+    it('should allow overriding commit message', async () => {
+      mockReleases(github, [
+        {
+          sha: 'abc123',
+          tagName: 'v1.0.0',
+          url: 'https://github.com/fake-owner/fake-repo/releases/tag/v1.0.0',
+        },
+      ]);
+      mockCommits(github, [
+        {
+          sha: 'def456',
+          message: 'fix: some bugfix',
+          files: [],
+          pullRequest: {
+            headBranchName: 'fix-1',
+            baseBranchName: 'main',
+            number: 123,
+            title: 'fix: some bugfix',
+            body: 'BEGIN_COMMIT_OVERRIDE\nfix: real fix message\nEND_COMMIT_OVERRIDE',
+            labels: [],
+            files: [],
+            sha: 'abc123',
+          },
+        },
+        {
+          sha: 'abc123',
+          message: 'chore: release 1.0.0',
+          files: [],
+          pullRequest: {
+            headBranchName: 'release-please/branches/main',
+            baseBranchName: 'main',
+            number: 123,
+            title: 'chore: release 1.0.0',
+            body: '',
+            labels: [],
+            files: [],
+            sha: 'abc123',
+          },
+        },
+      ]);
+      const manifest = new Manifest(
+        github,
+        'main',
+        {
+          '.': {
+            releaseType: 'simple',
+          },
+        },
+        {
+          '.': Version.parse('1.0.0'),
+        },
+        {
+          draftPullRequest: true,
+        }
+      );
+      const pullRequests = await manifest.buildPullRequests();
+      expect(pullRequests).lengthOf(1);
+      const pullRequest = pullRequests[0];
+      safeSnapshot(pullRequest.body.toString());
     });
 
     describe('with plugins', () => {
@@ -1831,6 +2042,69 @@ describe('Manifest', () => {
         sinon.assert.calledOnce(mockPlugin.run);
         sinon.assert.calledOnce(mockPlugin2.run);
       });
+    });
+
+    it('should fallback to tagged version', async () => {
+      mockReleases(github, []);
+      mockTags(github, [
+        {
+          name: 'pkg1-v1.0.0',
+          sha: 'abc123',
+        },
+      ]);
+      mockCommits(github, [
+        {
+          sha: 'def456',
+          message: 'fix: some bugfix',
+          files: [],
+        },
+        {
+          sha: 'abc123',
+          message: 'chore: release 1.0.0',
+          files: [],
+          pullRequest: {
+            headBranchName: 'release-please/branches/main/components/pkg1',
+            baseBranchName: 'main',
+            number: 123,
+            title: 'chore: release 1.0.0',
+            body: '',
+            labels: [],
+            files: [],
+            sha: 'abc123',
+          },
+        },
+      ]);
+      const getFileContentsStub = sandbox.stub(
+        github,
+        'getFileContentsOnBranch'
+      );
+      getFileContentsStub
+        .withArgs('package.json', 'main')
+        .resolves(
+          buildGitHubFileContent(
+            fixturesPath,
+            'manifest/repo/node/pkg1/package.json'
+          )
+        );
+      const manifest = new Manifest(
+        github,
+        'main',
+        {
+          '.': {
+            releaseType: 'node',
+          },
+        },
+        {
+          '.': Version.parse('1.0.0'),
+        }
+      );
+      const pullRequests = await manifest.buildPullRequests();
+      expect(pullRequests).lengthOf(1);
+      const pullRequest = pullRequests[0];
+      expect(pullRequest.version?.toString()).to.eql('1.0.1');
+      expect(pullRequest.headRefName).to.eql(
+        'release-please--branches--main--components--pkg1'
+      );
     });
   });
 
@@ -2954,6 +3228,41 @@ describe('Manifest', () => {
         .to.be.a('string')
         .and.satisfy((msg: string) => msg.startsWith('### [3.2.7]'));
       expect(releases[0].path).to.eql('.');
+    });
+
+    it('should skip component releases for non-component configs', async () => {
+      mockPullRequests(
+        github,
+        [],
+        [
+          {
+            headBranchName:
+              'release-please--branches--main--components--storage',
+            baseBranchName: 'main',
+            number: 1234,
+            title: 'chore(main): release storage 3.2.7',
+            body: pullRequestBody('release-notes/single.txt'),
+            labels: ['autorelease: pending'],
+            files: [],
+            sha: 'abc123',
+          },
+        ]
+      );
+      const manifest = new Manifest(
+        github,
+        'main',
+        {
+          '.': {
+            releaseType: 'simple',
+            includeComponentInTag: false,
+          },
+        },
+        {
+          '.': Version.parse('3.2.6'),
+        }
+      );
+      const releases = await manifest.buildReleases();
+      expect(releases).lengthOf(0);
     });
   });